refactor(vision): split planning and image rewriting out of the vision index (split S06 L1/2)
